Move findOrDefault in base DataSource model
