+++ packages/data-broker/src/DataBroker.js
@@ -8,7 +8,6 @@ import groupBy from 'lodash.groupby';
 import { ModelRegistry, TupaiaDatabase } from '@tupaia/database';
 import { countDistinct } from '@tupaia/utils';
 import { createService } from './services';
-import { modelClasses } from './modelClasses';
 
 let database;
 
@@ -21,7 +20,7 @@ const getDatabaseInstance = () => {
 
 export class DataBroker {
   constructor() {
-    this.models = new ModelRegistry(getDatabaseInstance(), modelClasses);
+    this.models = new ModelRegistry(getDatabaseInstance());
     this.resultMergers = {
       [this.getDataSourceTypes().DATA_ELEMENT]: this.mergeAnalytics,
       [this.getDataSourceTypes().DATA_GROUP]: this.mergeEvents,
