--- packages/data-broker/src/modelClasses/index.js
@@ -1,10 +1,0 @@
-/**
- * Tupaia
- * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
- */
-
-import { DataSourceModel as DataSource } from './DataSource';
-
-export const modelClasses = {
-  DataSource,
-};
