+++ packages/database/package.json
@@ -31,6 +31,7 @@
     "dotenv": "^8.2.0",
     "knex": "0.14.6",
     "lodash.clonedeep": "^4.5.0",
+    "lodash.keyby": "^4.6.0",
     "moment": "^2.24.0",
     "os": "0.1.1",
     "pg": "7.10.0",
